refactor: migrate coreops helpers into package
