+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+- Refactor: Introduced internal `c1c_coreops` package; legacy `shared/coreops_*` now deprecated shims. No behavior changes.
+
 ## v0.9.5 — 2025-10-22
 
 - Added audit reports under `AUDIT/20251022_COREOPS_AUDIT/`.
