+++ packages/c1c-coreops/README.md
@@ -1,6 +1,13 @@
 # c1c-coreops
 
-Reusable CoreOps helpers shared across C1C bots.
+Internal-only bundle of the CoreOps cog, RBAC helpers, embed renderers, and
+prefix detection utilities shared across C1C bots.
 
-This package is currently a placeholder while we extract the existing
-CoreOps logic from the recruitment bot into the shared workspace.
+## Contents
+- `c1c_coreops.cog` — CoreOps command handlers, telemetry views, and embeds.
+- `c1c_coreops.rbac` — Role gating checks aligned with the recruitment runtime.
+- `c1c_coreops.render` — Embed dataclasses and formatting helpers for digests.
+- `c1c_coreops.prefix` — Bang-command detection for admin shortcuts.
+
+Legacy modules continue to re-export these symbols from `shared/coreops_*` for
+one release so downstream importers have time to migrate.
