+++ packages/c1c-coreops/src/c1c_coreops/__init__.py
@@ -1,0 +1,28 @@
+"""CoreOps helpers packaged for internal reuse."""
+
+from . import cog as _cog
+from . import prefix as _prefix
+from . import rbac as _rbac
+from . import render as _render
+
+from .cog import *  # noqa: F401,F403
+from .prefix import *  # noqa: F401,F403
+from .rbac import *  # noqa: F401,F403
+from .render import *  # noqa: F401,F403
+
+__all__: list[str] = []
+_seen: set[str] = set()
+for _module in (_cog, _rbac, _render, _prefix):
+    names = getattr(_module, "__all__", None)
+    if names is None:
+        names = [name for name in vars(_module) if not name.startswith("_")]
+    for name in names:
+        if name in _seen:
+            continue
+        _seen.add(name)
+        __all__.append(name)
+
+__all__ = tuple(__all__)
+
+__docformat__ = "restructuredtext"
+
